# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RackScale API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  archived: "archived",
  draft: "draft",
} as const;

export interface Project {
  id: string;
  name: string;
  description?: string;
  location?: string;
  status: ProjectStatus;
  totalRackCount: number;
  scenarioCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectBodyStatus =
  (typeof CreateProjectBodyStatus)[keyof typeof CreateProjectBodyStatus];

export const CreateProjectBodyStatus = {
  active: "active",
  archived: "archived",
  draft: "draft",
} as const;

export interface CreateProjectBody {
  name: string;
  description?: string;
  location?: string;
  status?: CreateProjectBodyStatus;
  totalRackCount: number;
}

export type ScenarioCoolingType =
  (typeof ScenarioCoolingType)[keyof typeof ScenarioCoolingType];

export const ScenarioCoolingType = {
  air: "air",
  liquid: "liquid",
  hybrid: "hybrid",
} as const;

export type ScenarioRedundancyLevel =
  (typeof ScenarioRedundancyLevel)[keyof typeof ScenarioRedundancyLevel];

export const ScenarioRedundancyLevel = {
  N: "N",
  "N+1": "N+1",
  "2N": "2N",
} as const;

export interface Scenario {
  id: string;
  projectId: string;
  name: string;
  description?: string;
  rackCount: number;
  avgKwPerRack: number;
  coolingType: ScenarioCoolingType;
  redundancyLevel: ScenarioRedundancyLevel;
  pueTarget: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateScenarioBodyCoolingType =
  (typeof CreateScenarioBodyCoolingType)[keyof typeof CreateScenarioBodyCoolingType];

export const CreateScenarioBodyCoolingType = {
  air: "air",
  liquid: "liquid",
  hybrid: "hybrid",
} as const;

export type CreateScenarioBodyRedundancyLevel =
  (typeof CreateScenarioBodyRedundancyLevel)[keyof typeof CreateScenarioBodyRedundancyLevel];

export const CreateScenarioBodyRedundancyLevel = {
  N: "N",
  "N+1": "N+1",
  "2N": "2N",
} as const;

export interface CreateScenarioBody {
  name: string;
  description?: string;
  rackCount: number;
  avgKwPerRack: number;
  coolingType: CreateScenarioBodyCoolingType;
  redundancyLevel: CreateScenarioBodyRedundancyLevel;
  pueTarget: number;
}

export type RiskFlagLevel = (typeof RiskFlagLevel)[keyof typeof RiskFlagLevel];

export const RiskFlagLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type RiskFlagCategory =
  (typeof RiskFlagCategory)[keyof typeof RiskFlagCategory];

export const RiskFlagCategory = {
  power: "power",
  cooling: "cooling",
  capacity: "capacity",
  redundancy: "redundancy",
} as const;

export interface RiskFlag {
  level: RiskFlagLevel;
  category: RiskFlagCategory;
  message: string;
  recommendation: string;
}

export interface ScenarioResults {
  /** Total IT load in kW */
  totalItLoad: number;
  /** Total power draw including PUE in kW */
  totalPowerDraw: number;
  /** Required cooling capacity in kW */
  coolingCapacityRequired: number;
  /** Estimated annual energy cost in USD */
  estimatedAnnualCost: number;
  powerDensityKwPerSqft: number;
  co2EmissionsKgPerYear: number;
  riskFlags: RiskFlag[];
}

export type ScenarioWithResults = Scenario & {
  results: ScenarioResults;
};

export type ScenarioComparisonWinner = {
  scenarioId: string;
  reason: string;
};

export interface ScenarioComparison {
  projectId: string;
  scenarios: ScenarioWithResults[];
  winner: ScenarioComparisonWinner;
}

export interface DashboardStats {
  totalProjects: number;
  activeProjects: number;
  totalScenarios: number;
  totalItLoadKw: number;
  criticalRiskCount: number;
  avgPue: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  project_created: "project_created",
  scenario_created: "scenario_created",
  scenario_updated: "scenario_updated",
  risk_flagged: "risk_flagged",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  message: string;
  projectId: string;
  projectName: string;
  timestamp: string;
}

export interface RiskSummary {
  critical: number;
  high: number;
  medium: number;
  low: number;
  topRisks: RiskFlag[];
}

export type ListProjectsParams = {
  search?: string;
  status?: ListProjectsStatus;
};

export type ListProjectsStatus =
  (typeof ListProjectsStatus)[keyof typeof ListProjectsStatus];

export const ListProjectsStatus = {
  active: "active",
  archived: "archived",
  draft: "draft",
} as const;

export type CompareScenariosParams = {
  /**
   * Comma-separated scenario IDs
   */
  ids: string;
};