# lib/api-zod/src/generated/types/createScenarioBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RackScale API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateScenarioBodyCoolingType } from "./createScenarioBodyCoolingType";
import type { CreateScenarioBodyRedundancyLevel } from "./createScenarioBodyRedundancyLevel";

export interface CreateScenarioBody {
  name: string;
  description?: string;
  rackCount: number;
  avgKwPerRack: number;
  coolingType: CreateScenarioBodyCoolingType;
  redundancyLevel: CreateScenarioBodyRedundancyLevel;
  pueTarget: number;
}