# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RackScale API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalProjects: number;
  activeProjects: number;
  totalScenarios: number;
  totalItLoadKw: number;
  criticalRiskCount: number;
  avgPue: number;
}