# lib/api-zod/src/generated/types/scenarioResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RackScale API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFlag } from "./riskFlag";

export interface ScenarioResults {
  /** Total IT load in kW */
  totalItLoad: number;
  /** Total power draw including PUE in kW */
  totalPowerDraw: number;
  /** Required cooling capacity in kW */
  coolingCapacityRequired: number;
  /** Estimated annual energy cost in USD */
  estimatedAnnualCost: number;
  powerDensityKwPerSqft: number;
  co2EmissionsKgPerYear: number;
  riskFlags: RiskFlag[];
}